% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{determine_ambiguous_propositions}
\alias{determine_ambiguous_propositions}
\title{Determine which propositions were ambiguous based on norming responses.}
\usage{
determine_ambiguous_propositions()
}
\description{
Warning! Anything named `norms_responses` is clobbered from the env.
}

