% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/filters.R
\name{label_ambiguous_propositions}
\alias{label_ambiguous_propositions}
\title{Label the ambiguity of propositions based on norming data.}
\usage{
label_ambiguous_propositions(frame)
}
\description{
Label the ambiguity of propositions based on norming data.
}

