% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/filters.R
\name{label_bad_baseline_performance}
\alias{label_bad_baseline_performance}
\title{Label the propositions that were too hard based on baseline performance.}
\usage{
label_bad_baseline_performance(frame, question_first)
}
\description{
Label the propositions that were too hard based on baseline performance.
}

