% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_effects.R
\name{label_exposure_order}
\alias{label_exposure_order}
\title{Label the order that questions were exposed to participants.}
\usage{
label_exposure_order(frame)
}
\description{
Label the order that questions were exposed to participants.
}

