% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{label_outlier_map}
\alias{label_outlier_map}
\title{Label the outlier rows in a data.frame of subjects or items with reasons.}
\usage{
label_outlier_map(map, id_col, outlier_ids, reason,
  outlier_col = "is_outlier", reason_col = "reason")
}
\description{
Label the outlier rows in a data.frame of subjects or items with reasons.
}

