% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{label_outlier_props}
\alias{label_outlier_props}
\title{Partial wrapper around `label_outlier_map` for propositions.}
\usage{
label_outlier_props(props, prop_ids, reason)
}
\description{
Partial wrapper around `label_outlier_map` for propositions.
}

