% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{label_outlier_subjs}
\alias{label_outlier_subjs}
\title{Partial wrapper around `label_outlier_map` for subjects.}
\usage{
label_outlier_subjs(subjs, outlier_subjs, reason)
}
\description{
Partial wrapper around `label_outlier_map` for subjects.
}

