% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{label_outliers}
\alias{label_outliers}
\title{Label the outlier subjects and items in the data.}
\usage{
label_outliers(frame)
}
\description{
Outliers are labeled but no values are changed or dropped. To remove
outliers, filter them out.
}
\details{
> label_outliers(question_first) %>% filter(is_outlier == FALSE)
}

