% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/recoders.R
\name{recode_correct_response}
\alias{recode_correct_response}
\title{Recode the truth of the proposition.}
\usage{
recode_correct_response(frame)
}
\description{
Recode the truth of the proposition.
}

