% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/recoders.R
\name{recode_property_verification_data}
\alias{recode_property_verification_data}
\title{Recode variables for models and plots.}
\usage{
recode_property_verification_data(frame)
}
\description{
Recode variables for models and plots.
}

