% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recoders.R
\name{recode_trial_type}
\alias{recode_trial_type}
\title{Recode combination of feature and mask trial type.}
\usage{
recode_trial_type(frame)
}
\description{
Recode combination of feature and mask trial type.
}

