% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/filters.R
\name{determine_ambiguous_propositions}
\alias{determine_ambiguous_propositions}
\title{Determine which propositions were ambiguous.}
\usage{
determine_ambiguous_propositions()
}
\description{
Determine which propositions were ambiguous.
}

