% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/filters.R
\name{determine_bad_baseline_performance}
\alias{determine_bad_baseline_performance}
\title{Determine which propositions had bad baseline performance.}
\usage{
determine_bad_baseline_performance()
}
\description{
Determine which propositions had bad baseline performance.
}

