% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{label_outliers}
\alias{label_outliers}
\title{Label the outlier subjects and items in the data.}
\usage{
label_outliers(map, id_col, outlier_ids, reason, outlier_col = "is_outlier",
  reason_col = "reason")
}
\description{
Label the outlier subjects and items in the data.
}

